feature-policy/reporting/wakelock-reporting.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>

const check_report_format = (reports, observer) => {
  let report = reports[0];
  assert_equals(report.type, "feature-policy-violation");
  assert_equals(report.url, document.location.href);
  assert_equals(report.body.featureId, "wake-lock");
  assert_equals(report.body.sourceFile, document.location.href);
  assert_equals(typeof report.body.lineNumber, "number");
  assert_equals(typeof report.body.columnNumber, "number");
  assert_equals(report.body.disposition, "enforce");
};

promise_test(async t => {
  const report = new Promise(resolve => {
    new ReportingObserver((reports, observer) => resolve([reports, observer]),
                          {types: ['feature-policy-violation']}).observe();
  });

  await promise_rejects(t, 'NotAllowedError', WakeLock.request("screen"),
                        "WakeLock API should not be allowed in this document.");

  const [reports, observer] = await report;
  check_report_format(reports, observer);
}, "WakeLock Report Format");

</script>
